refactor: integrate gettext
This isn't actually useful yet, because I can't figure out how to use
xgettext properly with Zig files and can't be assed to try to manually
treat them as C files and merge the resulting pots. It's about time for
me to start zig-gettext... Anyways, going forward, translatable strings
will use gettext so in the future they can be translated.